+++ floship_llm/__init__.py
@@ -56,4 +56,4 @@ __all__ = [
     "lm_json_utils",
 ]
 
-__version__ = "0.5.54"
+__version__ = "0.6.0"